requirements.txt
asgiref==3.5.2
Babel==2.10.3
dj-database-url==1.0.0
Django==4.1.2
django-admin==2.0.1
django-crispy-forms==1.14.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-heroku==0.3.1
django-money==3.0.0
django-session-timeout==0.1.0
django-six==1.0.4
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.4
py-moneyed==2.0
pytz==2022.4
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.3.0
tzdata==2022.4
whitenoise==6.2.0
xlwt==1.3.0
